open_barrier/requirements.txt
python-telegram-bot
twilio
boto3-stubs[dynamodb]
boto3
